requirements.txt
appdirs==1.4.2
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
Django==1.10.6
django-appconf==1.0.2
django-classy-tags==0.8.0
django-crispy-forms==1.6.1
djangotoolbox==1.8.0
enum34==1.1.6
futures==3.2.0
geopy==1.11.0
google-api-core==1.0.0
google-auth==1.4.1
google-cloud-language==1.0.1
googleapis-common-protos==1.5.3
grpcio==1.10.0
idna==2.6
oauthlib==2.0.1
packaging==16.8
pkg-resources==0.0.0
protobuf==3.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pymongo==3.4.0
pyparsing==2.2.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
tweepy==3.5.0
urllib3==1.22
